added  @Fetch(FetchMode.SUBSELECT) to Contestant to eliminate JPA exceptions

+++ src/main/java/com/github/pawelbialas/testgeneratorapp/entity/result/dto/ResultDto.java
@@ -19,7 +19,7 @@ import java.util.UUID;
 public class ResultDto extends BaseItem {
 
 
-    private Integer score;
+    private Float score;
     private SkillTestDto skillTest;
     private ContestantDto contestant;
 
@@ -28,7 +28,7 @@ public class ResultDto extends BaseItem {
                      Long version,
                      OffsetDateTime createdDate,
                      OffsetDateTime lastModifiedDate,
-                     Integer score,
+                     Float score,
                      SkillTestDto skillTest,
                      ContestantDto contestant
     ) {
